+++ src/test/integration/linearCommands.test.ts
@@ -3,12 +3,12 @@
  */
 
 // Setup must be imported FIRST to register mocks before extension activation
-import "./setup"
+import "../setup"
 
 import * as assert from "assert"
 import * as vscode from "vscode"
-import {createTestDocument, closeAllEditors} from "./testUtils"
-import {TEST_SLACK_URLS, TEST_LINEAR_ISSUES} from "./fixtures"
+import {createTestDocument, closeAllEditors} from "../testUtils"
+import {TEST_SLACK_URLS, TEST_LINEAR_ISSUES} from "../fixtures"
 
 suite("Linear Commands Tests", () => {
   suiteSetup(async () => {